requirements.txt
flask==3.0.0
wtforms==3.1.1
argon2-cffi==23.1.0
pyserial==3.5
flask-socketio==5.3.6
matplotlib==3.9.0
opencv-python==4.9.0.80
scipy==1.13.1
peakutils==1.3.4
pandas==2.2.2